Implement bot configuration and user management features with Pattern designs and dependency injection
