+++ requirements.txt
@@ -39,6 +39,7 @@ pydantic==2.11.3
 pydantic_core==2.33.1
 python-dateutil==2.9.0.post0
 python-telegram-bot==22.0
+pytz==2025.2
 requests==2.32.3
 rsa==4.9
 six==1.17.0
